refactor: change endpoints to app router
