+++ src/proyectos/index.ts
@@ -1,3 +1,4 @@
+export * from "./api"
 export * from "./button"
 export * from "./cards"
 export * from "./form"
